Fix Dijkstra's

+++ src/grapes/cgraph/heap.h
@@ -11,7 +11,7 @@ typedef struct MinHeap_s MinHeap;
 
 MinHeap   *MinHeap_alloc(Py_ssize_t max_size);
 void       MinHeap_free(MinHeap *heap);
-void       MinHeap_insert(MinHeap *heap, Py_ssize_t key, Py_ssize_t priority);
+void       MinHeap_insert(MinHeap *heap, Py_ssize_t key, double priority);
 Py_ssize_t MinHeap_extract_min(MinHeap *heap);
 int        MinHeap_is_empty(MinHeap *heap);
 void MinHeap_siftdown(MinHeap *heap, Py_ssize_t startpos, Py_ssize_t pos);
